Convert create volunteer family dialog to drawer
